Rename main search component; combine results components
